[ot] python/qemu: ot.verilator: move VCP management
Move it to a dedicated class
